Share the run-or-throw command helper
The publish gate and the tool-load smoke each defined a byte-identical
helper that spawns a command and throws its captured output on a
non-zero exit, differing only in the error prefix. A shared factory now
builds one from a prefix; both consume it.
